tests: fix client integration test
