--- tests/integration/ping-pong/.eslintrc
@@ -1,3 +1,0 @@
-{
-  "extends": "@ircam",
-}
