--- tests/integration/ping-pong/.npmrc
@@ -1,1 +1,0 @@
-package-lock=false
