--- tests/integration/ping-pong/.soundworks
@@ -1,5 +1,0 @@
-{
-  "name": "ping-pong",
-  "eslint": true,
-  "language": "js"
-}
