--- tests/integration/ping-pong/config/application.json
@@ -1,10 +1,0 @@
-{
-  name: 'ping-pong',
-  author: '',
-  clients: {
-    player: {
-      target: 'browser',
-      default: true,
-    },
-  },
-}
