--- tests/integration/ping-pong/src/clients/player/index.js
@@ -1,45 +1,0 @@
-import '@soundworks/helpers/polyfills.js';
-import Client from '../../../../../../src/client/Client.js';
-import launcher from '@soundworks/helpers/launcher.js';
-
-import { html, render } from 'lit';
-import '../components/sw-audit.js';
-
-// - General documentation: https://soundworks.dev/
-// - API documentation:     https://soundworks.dev/api
-// - Issue Tracker:         https://github.com/collective-soundworks/soundworks/issues
-// - Wizard & Tools:        `npx soundworks`
-
-const config = window.SOUNDWORKS_CONFIG;
-
-async function main($container) {
-  const client = new Client(config);
-
-  launcher.register(client, {
-    initScreensContainer: $container,
-    reloadOnVisibilityChange: false,
-  });
-
-  await client.start();
-
-  function renderApp() {
-    render(html`
-      <div class="controller-layout">
-        <header>
-          <h1>${client.config.app.name} | ${client.role}</h1>
-          <sw-audit .client="${client}"></sw-audit>
-        </header>
-        <section>
-          <p>Hello ${client.config.app.name}!</p>
-        </section>
-      </div>
-    `, $container);
-  }
-
-  renderApp();
-}
-
-launcher.execute(main, {
-  numClients: parseInt(new URLSearchParams(window.location.search).get('emulate')) || 1,
-  width: '50%',
-});
