--- tests/integration/ping-pong/src/server/schemas/.gitkeep
@@ -1,1 +1,0 @@
-.gitkeep
