--- tests/integration/ping-pong/src/utils/catch-unhandled-errors.js
@@ -1,9 +1,0 @@
-process.on('unhandledRejection', (reason, _p) => {
-  console.log('> Unhandled Promise Rejection');
-  console.log(reason);
-});
-
-process.on('uncaughtException', (reason, _p) => {
-  console.log('> Unhandled Promise Rejection');
-  console.log(reason);
-});
