--- tests/integration/test-browser/.eslintrc
@@ -1,3 +1,0 @@
-{
-  "extends": "@ircam",
-}
