--- tests/integration/test-browser/config/application.json
@@ -1,10 +1,0 @@
-{
-  name: 'browser-integration',
-  author: '',
-  clients: {
-    player: {
-      target: 'browser',
-      default: true,
-    },
-  },
-}
