+++ tests/integration/test-browser/config/application.yaml
@@ -1,0 +1,6 @@
+name: test-browser
+author: ""
+clients:
+  player:
+    runtime: browser
+    default: true
