+++ tests/integration/test-browser/eslint.config.js
@@ -1,0 +1,1 @@
+export { default } from '@ircam/eslint-config';
