+++ tests/integration/test-browser/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "browser-integration",
+  "name": "test-browser",
   "description": "soundworks application",
   "authors": [],
   "license": "BSD-3-Clause",
@@ -7,29 +7,27 @@
   "type": "module",
   "private": true,
   "scripts": {
-    "build": "npm run clean && sass src/clients/components/css:.build/public/css && soundworks-build -b",
-    "build:production": "npm run clean && soundworks-build -b -m",
+    "build": "npm run clean && sass src/styles:.build/public/css && soundworks-build -b",
     "clean": "soundworks-build -D",
     "dev": "npm run build && (concurrently -i -p \"none\" \"npm run watch:inspect server\" \"soundworks-build -b -w\" \"npm run watch:sass\")",
-    "postinstall": "soundworks-build -C",
+    "lint": "npx eslint src",
     "start": "node .build/server/index.js",
     "watch": "soundworks-build -p",
-    "watch:inspect": "soundworks-build -d -p",
-    "watch:sass": "sass --watch src/clients/components/css:.build/public/css",
-    "lint": "eslint ."
+    "watch:inspect": "soundworks-build -i -p",
+    "watch:sass": "sass --watch src/styles:.build/public/css"
   },
   "dependencies": {
-    "@soundworks/core": "^4.0.0-alpha.0",
-    "@soundworks/helpers": "^1.0.0-alpha.2",
-    "json5": "^2.2.2",
-    "lit": "^2.2.2"
+    "@soundworks/helpers": "^5.0.0",
+    "lit": "^3.1.3"
   },
   "devDependencies": {
-    "@ircam/eslint-config": "^1.3.0",
-    "@soundworks/build": "^1.0.0-alpha.0",
-    "@soundworks/create": "^1.0.0-alpha.12",
-    "concurrently": "^7.1.0",
-    "eslint": "^8.32.0",
-    "sass": "^1.57.1"
+    "@ircam/eslint-config": "^2.0.1",
+    "@soundworks/build": "^5.0.0",
+    "@soundworks/create": "^6.0.1",
+    "@types/node": "^22.10.10",
+    "@types/web": "^0.0.199",
+    "concurrently": "^8.2.2",
+    "eslint": "^9.18.0",
+    "sass": "^1.77.1"
   }
 }
