--- tests/integration/test-browser/src/clients/components/css/app.scss
@@ -1,31 +1,0 @@
-:root {
-  --sw-background-color: #000000;
-  --sw-light-background-color: #242424;
-  --sw-font-color: #ffffff;
-  --sw-light-font-color: #cccccc;
-  --sw-font-color-error: #a94442;
-  --sw-font-family: Consolas, monaco, monospace;
-  --sw-font-size: 62.5%; // such that 1rem == 10px
-}
-
-html, body {
-  width: 100%;
-  min-height: 100vh;
-  background-color: var(--sw-background-color);
-  color: var(--sw-font-color);
-}
-
-html {
-  font-size: var(--sw-font-size);
-}
-
-* {
-  box-sizing: border-box;
-  font-family: var(--sw-font-family);
-}
-
-body.loading {
-  background-image: url(../../images/loader.gif);
-  background-position: 50% 50%;
-  background-repeat: no-repeat;
-}
