+++ tests/integration/test-browser/src/clients/player.js
@@ -1,0 +1,31 @@
+import '@soundworks/helpers/polyfills.js';
+import { loadConfig, launcher } from '@soundworks/helpers/browser.js';
+
+import { Client } from '../../../../../src/client/index.js';
+
+// - General documentation: https://soundworks.dev/
+// - API documentation:     https://soundworks.dev/api
+// - Issue Tracker:         https://github.com/collective-soundworks/soundworks/issues
+// - Wizard & Tools:        `npx soundworks`
+
+async function main($container) {
+  const config = loadConfig();
+  const client = new Client(config);
+
+  // Eventually register plugins
+  // client.pluginManager.register('my-plugin', plugin);
+
+  // cf. https://soundworks.dev/tools/helpers.html#browserlauncher
+  launcher.register(client, { initScreensContainer: $container });
+
+  await client.start();
+
+  const globals = await client.stateManager.attach('globals');
+  globals.set({ done: true });
+}
+
+// The launcher allows to launch multiple clients in the same browser window
+// e.g. `http://127.0.0.1:8000?emulate=10` to run 10 clients side-by-side
+launcher.execute(main, {
+  numClients: parseInt(new URLSearchParams(window.location.search).get('emulate') || '') || 1,
+});
