--- tests/integration/test-browser/src/server/schemas/.gitkeep
@@ -1,1 +1,0 @@
-.gitkeep
