--- tests/integration/test-browser/src/server/tmpl/default.tmpl
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no, shrink-to-fit=no">
-    <meta name="theme-color" content="#000000">
-
-    <title>
-      ${d.app.name} | ${d.role}
-    </title>
-
-    <script>
-      window.SOUNDWORKS_CONFIG = JSON.parse('${JSON.stringify(d)}');
-    </script>
-
-    ${
-      (d.env.type === 'production')
-        ? `<script defer src="${d.env.subpath ? `/${d.env.subpath}` : ''}/build/${d.role}.min.js"></script>`
-        : `<script defer src="${d.env.subpath ? `/${d.env.subpath}` : ''}/build/${d.role}.js"></script>`
-    }
-
-    <link rel="stylesheet" href="${d.env.subpath ? `/${d.env.subpath}` : ''}/build/css/normalize.css">
-    <link rel="stylesheet" href="${d.env.subpath ? `/${d.env.subpath}` : ''}/build/css/app.css">
-  </head>
-
-  <body class="loading ${d.role}"></body>
-</html>
